Fase front

+++ ProductAPI/Data/Repositories/ICategoryRepository.cs
@@ -5,7 +5,7 @@ namespace ProductAPI.Data.Repositories
     public interface ICategoryRepository
     {
         Task<IEnumerable<Category>> GetAllAsync();
-        Task<Category> GetByIdAsync(int id);
+        Task<Category> GetByIdAsync(long id);
         Task AddAsync(Category category);
         void Update(Category category);
         void Delete(Category category);
